graphql: Updated the README.
Made the todo list a little more consistent, adding a link to nvm.
Typos and such.
More details on how to get the Bearer token (and advice on how to avoid my mistake of grabbing a truncated version of the token the first time I tried).
Added a simple example query just to get the ball rolling.