web: added edit engagement button to engagement details page
I also added a refetchQueries prop to the EditEngagementModal.
Though this solution is slightly better than directly importing
names into the Modal, ideally this modal should need to know
query names. There's a problem in the way Apollo is determing
which queries are active. Apollo should not need to be told
which queries to rerun. There's a task to look into this issue
further.

https://app.asana.com/0/1202198978668199/1202512448907854/f